+++ R/is_url.R
@@ -6,7 +6,6 @@
 #' Borrowed from \code{seqminer} internal function}
 #' 
 #' @keywords internal
-#' @importFrom RCurl url.exists
 is_url <- function(path,
                    protocols=c("http","https",
                                "ftp","ftps",
@@ -17,6 +16,7 @@ is_url <- function(path,
     pattern <- paste(paste0("^",protocols,"://"),collapse = "|")
     if (grepl(pattern = pattern, x = path, ignore.case = TRUE)) {
         if(isTRUE(check_exists)){
+            requireNamespace("RCurl")
             return(RCurl::url.exists(path))
         }
         return(TRUE)